requirements.txt
django>=2.2.10
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.8.4,<2.9.0
django-phone-field
flake8>=3.6.0,<3.7.0
twilio==6.39.0
